+++ tt_metal/fabric/hw/inc/edm_fabric/fabric_edm_packet_header_validate.hpp
@@ -4,7 +4,7 @@
 
 #pragma once
 
-#include "cpp/ttnn/operations/ccl/kernels/edm_fabric/fabric_edm_packet_header.hpp"
+#include "fabric_edm_packet_header.hpp"
 #include "debug/assert.h"
 
 namespace tt::fabric {
@@ -12,7 +12,7 @@ namespace tt::fabric {
 FORCE_INLINE void validate(const PacketHeader& packet_header) {
     ASSERT(packet_header.chip_send_type <= CHIP_SEND_TYPE_LAST);
 }
-FORCE_INLINE bool is_valid(PacketHeader const& packet_header) {
+FORCE_INLINE bool is_valid(const PacketHeader& packet_header) {
     return (packet_header.chip_send_type <= CHIP_SEND_TYPE_LAST) && (packet_header.noc_send_type <= NOC_SEND_TYPE_LAST);
 }
