+++ tt_metal/fabric/hw/inc/edm_fabric/fabric_erisc_datamover_channels.hpp
@@ -13,13 +13,12 @@
 #include "tt_metal/hw/inc/ethernet/tunneling.h"
 #include "tt_metal/hw/inc/utils/utils.h"
 #include "risc_attribs.h"
-#include "cpp/ttnn/operations/ccl/kernels/edm_fabric/fabric_edm_packet_header.hpp"
-#include "cpp/ttnn/operations/ccl/kernels/edm_fabric/fabric_edm_types.hpp"
-#include "cpp/ttnn/operations/ccl/shared_with_host/hetergeneous_data_structs.hpp"
-#include "ttnn/cpp/ttnn/operations/ccl/kernels/edm_fabric/edm_fabric_worker_adapters.hpp"
-#include "cpp/ttnn/operations/ccl/kernels/edm_fabric/edm_fabric_flow_control_helpers.hpp"
-namespace tt::fabric {
+#include "fabric_edm_packet_header.hpp"
+#include "fabric_edm_types.hpp"
+#include "edm_fabric_worker_adapters.hpp"
+#include "edm_fabric_flow_control_helpers.hpp"
 
+namespace tt::fabric {
 
 template <typename T>
 FORCE_INLINE auto wrap_increment(T val, size_t max) {
@@ -28,7 +27,7 @@ FORCE_INLINE auto wrap_increment(T val, size_t max) {
 
 template <uint8_t NUM_BUFFERS>
 class EthChannelBuffer final {
-   public:
+public:
     // The channel structure is as follows:
     //              &header->  |----------------| channel_base_address
     //                         |    header      |
@@ -55,54 +54,45 @@ class EthChannelBuffer final {
         max_eth_payload_size_in_bytes(buffer_size_in_bytes + sizeof(eth_channel_sync_t)),
         channel_id(channel_id) {
         for (uint8_t i = 0; i < NUM_BUFFERS; i++) {
-            this->buffer_addresses[i] =
-                channel_base_address + i * this->max_eth_payload_size_in_bytes;
+            this->buffer_addresses[i] = channel_base_address + i * this->max_eth_payload_size_in_bytes;
         }
     }
 
-    [[nodiscard]] FORCE_INLINE size_t get_buffer_address(BufferIndex const& buffer_index) const {
+    [[nodiscard]] FORCE_INLINE size_t get_buffer_address(const BufferIndex& buffer_index) const {
         return this->buffer_addresses[buffer_index];
     }
 
     template <typename T>
-    [[nodiscard]] FORCE_INLINE volatile T *get_packet_header(BufferIndex const& buffer_index) const {
-        return reinterpret_cast<volatile T *>(this->buffer_addresses[buffer_index]);
+    [[nodiscard]] FORCE_INLINE volatile T* get_packet_header(const BufferIndex& buffer_index) const {
+        return reinterpret_cast<volatile T*>(this->buffer_addresses[buffer_index]);
     }
 
     template <typename T>
-    [[nodiscard]] FORCE_INLINE size_t get_payload_size(BufferIndex const& buffer_index) const {
+    [[nodiscard]] FORCE_INLINE size_t get_payload_size(const BufferIndex& buffer_index) const {
         return get_packet_header<T>(buffer_index)->get_payload_size_including_header();
     }
-    [[nodiscard]] FORCE_INLINE size_t get_channel_buffer_max_size_in_bytes(BufferIndex const& buffer_index) const {
+    [[nodiscard]] FORCE_INLINE size_t get_channel_buffer_max_size_in_bytes(const BufferIndex& buffer_index) const {
         return this->buffer_size_in_bytes;
     }
 
     // Doesn't return the message size, only the maximum eth payload size
-    [[nodiscard]] FORCE_INLINE size_t get_max_eth_payload_size() const {
-        return this->max_eth_payload_size_in_bytes;
-    }
+    [[nodiscard]] FORCE_INLINE size_t get_max_eth_payload_size() const { return this->max_eth_payload_size_in_bytes; }
 
     [[nodiscard]] FORCE_INLINE size_t get_id() const { return this->channel_id; }
 
-    [[nodiscard]] FORCE_INLINE bool eth_is_acked_or_completed(BufferIndex const& buffer_index) const {
+    [[nodiscard]] FORCE_INLINE bool eth_is_acked_or_completed(const BufferIndex& buffer_index) const {
         return eth_is_receiver_channel_send_acked(buffer_index) || eth_is_receiver_channel_send_done(buffer_index);
     }
 
-
-    FORCE_INLINE bool needs_to_send_channel_sync() const {
-        return this->need_to_send_channel_sync;
-    }
+    FORCE_INLINE bool needs_to_send_channel_sync() const { return this->need_to_send_channel_sync; }
 
     FORCE_INLINE void set_need_to_send_channel_sync(bool need_to_send_channel_sync) {
         this->need_to_send_channel_sync = need_to_send_channel_sync;
     }
 
-    FORCE_INLINE void clear_need_to_send_channel_sync() {
-        this->need_to_send_channel_sync = false;
-    }
+    FORCE_INLINE void clear_need_to_send_channel_sync() { this->need_to_send_channel_sync = false; }
 
-   private:
-
+private:
     std::array<size_t, NUM_BUFFERS> buffer_addresses;
 
     // header + payload regions only
@@ -111,7 +101,6 @@ class EthChannelBuffer final {
     const std::size_t max_eth_payload_size_in_bytes;
     uint8_t channel_id;
 };
-
 
 template <uint8_t NUM_BUFFERS>
 struct EdmChannelWorkerInterface {
@@ -131,9 +120,9 @@ struct EdmChannelWorkerInterface {
         // completion field so that way we don't have to do a volatile read for every
         // packet... Then we'll also be able to cache the uint64_t addr of the worker
         // semaphore directly (saving on regenerating it each time)
-        volatile EDMChannelWorkerLocationInfo *worker_location_info_ptr,
-        volatile tt_l1_ptr uint32_t *const remote_producer_wrptr,
-        volatile tt_l1_ptr uint32_t *const connection_live_semaphore) :
+        volatile EDMChannelWorkerLocationInfo* worker_location_info_ptr,
+        volatile tt_l1_ptr uint32_t* const remote_producer_wrptr,
+        volatile tt_l1_ptr uint32_t* const connection_live_semaphore) :
         worker_location_info_ptr(worker_location_info_ptr),
         cached_worker_semaphore_address(0),
         remote_producer_wrptr(remote_producer_wrptr),
@@ -141,20 +130,17 @@ struct EdmChannelWorkerInterface {
         local_wrptr(),
         local_ackptr(),
         local_rdptr() {
-        DPRINT << "EDM  my_x: " << (uint32_t)my_x[0] << ", my_y: " << (uint32_t)my_y[0] << " rdptr set to 0 at " << (uint32_t)(void*)&(worker_location_info_ptr->edm_rdptr) << "\n";
+        DPRINT << "EDM  my_x: " << (uint32_t)my_x[0] << ", my_y: " << (uint32_t)my_y[0] << " rdptr set to 0 at "
+               << (uint32_t)(void*)&(worker_location_info_ptr->edm_rdptr) << "\n";
         *reinterpret_cast<volatile uint32_t*>(&(worker_location_info_ptr->edm_rdptr)) = 0;
-        }
+    }
 
     // Flow control methods
     //
     // local_wrptr trails from_remote_wrptr
     // we have new data if they aren't equal
-    [[nodiscard]] FORCE_INLINE bool has_unsent_payload() {
-        return local_wrptr.get_ptr() != *remote_producer_wrptr;
-    }
-    [[nodiscard]] FORCE_INLINE bool has_unacked_sends() {
-        return local_ackptr.get_ptr() != local_wrptr.get_ptr();
-    }
+    [[nodiscard]] FORCE_INLINE bool has_unsent_payload() { return local_wrptr.get_ptr() != *remote_producer_wrptr; }
+    [[nodiscard]] FORCE_INLINE bool has_unacked_sends() { return local_ackptr.get_ptr() != local_wrptr.get_ptr(); }
 
     [[nodiscard]] FORCE_INLINE uint32_t get_worker_semaphore_address() const {
         return cached_worker_semaphore_address & 0xFFFFFFFF;
@@ -167,9 +153,11 @@ struct EdmChannelWorkerInterface {
     // Connection management methods
     //
     FORCE_INLINE void teardown_connection(uint32_t last_edm_rdptr_value) const {
-        auto const &worker_info = *worker_location_info_ptr;
+        const auto& worker_info = *worker_location_info_ptr;
         uint64_t worker_semaphore_address = get_noc_addr(
-            (uint32_t)worker_info.worker_xy.x, (uint32_t)worker_info.worker_xy.y, worker_info.worker_teardown_semaphore_address);
+            (uint32_t)worker_info.worker_xy.x,
+            (uint32_t)worker_info.worker_xy.y,
+            worker_info.worker_teardown_semaphore_address);
 
         // Set connection to unused so it's available for next worker
         *this->connection_live_semaphore = tt::fabric::EdmToEdmSender<0>::unused_connection_value;
@@ -180,33 +168,30 @@ struct EdmChannelWorkerInterface {
     }
 
     FORCE_INLINE void cache_producer_noc_addr() {
-        auto const &worker_info = *worker_location_info_ptr;
+        const auto& worker_info = *worker_location_info_ptr;
         uint64_t worker_semaphore_address = get_noc_addr(
-            (uint32_t)worker_info.worker_xy.x,
-            (uint32_t)worker_info.worker_xy.y,
-            worker_info.worker_semaphore_address);
+            (uint32_t)worker_info.worker_xy.x, (uint32_t)worker_info.worker_xy.y, worker_info.worker_semaphore_address);
         this->cached_worker_semaphore_address = worker_semaphore_address;
     }
 
-    FORCE_INLINE bool all_eth_packets_acked() const {
-        return this->local_ackptr.is_caught_up_to(this->local_wrptr);
+    FORCE_INLINE bool all_eth_packets_acked() const { return this->local_ackptr.is_caught_up_to(this->local_wrptr); }
+    FORCE_INLINE bool all_eth_packets_completed() const { return this->local_rdptr.is_caught_up_to(this->local_wrptr); }
+
+    [[nodiscard]] FORCE_INLINE bool has_worker_teardown_request() const {
+        return *connection_live_semaphore == tt::fabric::EdmToEdmSender<0>::close_connection_request_value;
     }
-    FORCE_INLINE bool all_eth_packets_completed() const {
-        return this->local_rdptr.is_caught_up_to(this->local_wrptr);
+    [[nodiscard]] FORCE_INLINE bool connection_is_live() const {
+        return *connection_live_semaphore == tt::fabric::EdmToEdmSender<0>::open_connection_value;
     }
 
-    [[nodiscard]] FORCE_INLINE bool has_worker_teardown_request() const { return *connection_live_semaphore == tt::fabric::EdmToEdmSender<0>::close_connection_request_value; }
-    [[nodiscard]] FORCE_INLINE bool connection_is_live() const { return *connection_live_semaphore == tt::fabric::EdmToEdmSender<0>::open_connection_value; }
-
-    volatile EDMChannelWorkerLocationInfo *worker_location_info_ptr;
+    volatile EDMChannelWorkerLocationInfo* worker_location_info_ptr;
     uint64_t cached_worker_semaphore_address = 0;
-    volatile tt_l1_ptr uint32_t *const remote_producer_wrptr;
-    volatile tt_l1_ptr uint32_t *const connection_live_semaphore;
+    volatile tt_l1_ptr uint32_t* const remote_producer_wrptr;
+    volatile tt_l1_ptr uint32_t* const connection_live_semaphore;
 
     ChannelBufferPointer<NUM_BUFFERS> local_wrptr;
     ChannelBufferPointer<NUM_BUFFERS> local_ackptr;
-    ChannelBufferPointer<NUM_BUFFERS> local_rdptr; // also used as completion_ptr
+    ChannelBufferPointer<NUM_BUFFERS> local_rdptr;  // also used as completion_ptr
 };
-
 
 }  // namespace tt::fabric
