+++ tt_metal/fabric/hw/inc/edm_fabric/worker_edm_utils.hpp
@@ -6,29 +6,17 @@
 
 #include "dataflow_api.h"
 #include "debug/assert.h"
-#include "cpp/ttnn/operations/ccl/shared_with_host/hetergeneous_data_structs.hpp"
+#include "tt_metal/api/tt-metalium/fabric_edm_types.hpp"
 
-using ttnn::ccl::ShardType;
-using ttnn::ccl::WorkerXY;
-// using ttnn::ccl::coord_t;
+namespace tt::fabric {
 
-namespace ttnn {
-namespace ccl {
 static FORCE_INLINE coord_t coord_from_args(std::size_t& arg_idx) {
     uint32_t x = get_arg_val<uint32_t>(arg_idx++);
     uint32_t y = get_arg_val<uint32_t>(arg_idx++);
     return coord_t(x, y);
 }
 
-enum EDM_IO_BLOCKING_MODE {
-    FLUSH_BLOCKING,
-    BLOCKING,
-    NON_BLOCKING
-};
-
-}  // namespace ccl
-}  // namespace ttnn
-
+enum EDM_IO_BLOCKING_MODE { FLUSH_BLOCKING, BLOCKING, NON_BLOCKING };
 
 FORCE_INLINE void push_filler_pages_to_cb(const uint32_t& cb_id, uint32_t num_pages) {
     ASSERT(num_pages < get_local_cb_interface(cb_id).fifo_num_pages);
@@ -50,40 +38,51 @@ FORCE_INLINE void fetch_chunk(
     cb_push_back(cb_id, num_pages);
 }
 
-template<ttnn::ccl::EDM_IO_BLOCKING_MODE blocking_mode = ttnn::ccl::EDM_IO_BLOCKING_MODE::BLOCKING>
+template <EDM_IO_BLOCKING_MODE blocking_mode = EDM_IO_BLOCKING_MODE::BLOCKING>
 FORCE_INLINE void send_chunk_from_address_with_trid(
-    const uint32_t& local_l1_address, const uint32_t& num_pages, const uint32_t& page_size, uint32_t remote_l1_write_addr, uint8_t trid, uint8_t cmd_buf) {
-    noc_async_write_one_packet_with_trid_with_state(local_l1_address, remote_l1_write_addr, page_size * num_pages, trid, cmd_buf);
+    const uint32_t& local_l1_address,
+    const uint32_t& num_pages,
+    const uint32_t& page_size,
+    uint32_t remote_l1_write_addr,
+    uint8_t trid,
+    uint8_t cmd_buf) {
+    noc_async_write_one_packet_with_trid_with_state(
+        local_l1_address, remote_l1_write_addr, page_size * num_pages, trid, cmd_buf);
     // TODO: this barrier will no longer be functional since we are not incrementing noc counters, remove
-    if constexpr (blocking_mode == ttnn::ccl::EDM_IO_BLOCKING_MODE::FLUSH_BLOCKING) {
+    if constexpr (blocking_mode == EDM_IO_BLOCKING_MODE::FLUSH_BLOCKING) {
         noc_async_writes_flushed();
-    } else if constexpr (blocking_mode == ttnn::ccl::EDM_IO_BLOCKING_MODE::BLOCKING) {
+    } else if constexpr (blocking_mode == EDM_IO_BLOCKING_MODE::BLOCKING) {
         noc_async_write_barrier();
     }
 }
 
-template<ttnn::ccl::EDM_IO_BLOCKING_MODE blocking_mode = ttnn::ccl::EDM_IO_BLOCKING_MODE::BLOCKING>
+template <EDM_IO_BLOCKING_MODE blocking_mode = EDM_IO_BLOCKING_MODE::BLOCKING>
 FORCE_INLINE void send_chunk_from_address(
-    const uint32_t& local_l1_address, const uint32_t& num_pages, const uint32_t& page_size, uint64_t remote_l1_write_addr) {
+    const uint32_t& local_l1_address,
+    const uint32_t& num_pages,
+    const uint32_t& page_size,
+    uint64_t remote_l1_write_addr) {
     noc_async_write(local_l1_address, remote_l1_write_addr, page_size * num_pages);
-    if constexpr (blocking_mode == ttnn::ccl::EDM_IO_BLOCKING_MODE::FLUSH_BLOCKING) {
+    if constexpr (blocking_mode == EDM_IO_BLOCKING_MODE::FLUSH_BLOCKING) {
         noc_async_writes_flushed();
-    } else if constexpr (blocking_mode == ttnn::ccl::EDM_IO_BLOCKING_MODE::BLOCKING) {
+    } else if constexpr (blocking_mode == EDM_IO_BLOCKING_MODE::BLOCKING) {
         noc_async_write_barrier();
     }
 }
 
-template<ttnn::ccl::EDM_IO_BLOCKING_MODE blocking_mode = ttnn::ccl::EDM_IO_BLOCKING_MODE::BLOCKING>
+template <EDM_IO_BLOCKING_MODE blocking_mode = EDM_IO_BLOCKING_MODE::BLOCKING>
 FORCE_INLINE void send_chunk(
     const uint32_t& cb_id, const uint32_t& num_pages, const uint32_t& page_size, uint64_t remote_l1_write_addr) {
     cb_wait_front(cb_id, num_pages);
     uint32_t l1_read_addr = get_read_ptr(cb_id);
     noc_async_write(l1_read_addr, remote_l1_write_addr, page_size * num_pages);
-    if constexpr (blocking_mode == ttnn::ccl::EDM_IO_BLOCKING_MODE::FLUSH_BLOCKING) {
+    if constexpr (blocking_mode == EDM_IO_BLOCKING_MODE::FLUSH_BLOCKING) {
         noc_async_writes_flushed();
         cb_pop_front(cb_id, num_pages);
-    } else if constexpr (blocking_mode == ttnn::ccl::EDM_IO_BLOCKING_MODE::BLOCKING) {
+    } else if constexpr (blocking_mode == EDM_IO_BLOCKING_MODE::BLOCKING) {
         noc_async_write_barrier();
         cb_pop_front(cb_id, num_pages);
     }
 }
+
+}  // namespace tt::fabric
