+++ ttnn/cpp/ttnn/operations/ccl/ccl_pybind.cpp
@@ -9,7 +9,7 @@
 #include "ttnn/operations/ccl/barrier/barrier_pybind.hpp"
 
 #include "ttnn/operations/ccl/ccl_host_datastructures.hpp"
-#include "cpp/ttnn/operations/ccl/erisc_datamover_builder.hpp"
+#include "cpp/ttnn/operations/ccl/erisc_datamover_builder_helper.hpp"
 
 namespace ttnn::operations::ccl {
