+++ ttnn/cpp/ttnn/operations/experimental/ccl/reduce_scatter_async/device/reduce_scatter_async_program.cpp
@@ -14,7 +14,7 @@
 #include <tt-metalium/device_impl.hpp>
 #include <tt-metalium/kernel_types.hpp>
 #include <tt-metalium/span.hpp>
-#include "cpp/ttnn/operations/ccl/erisc_datamover_builder.hpp"
+#include <tt-metalium/erisc_datamover_builder.hpp>
 #include "cpp/ttnn/operations/ccl/common/host/ccl_worker_builder.hpp"
 #include <tt-metalium/host_api.hpp>
 #include "ttnn/operation.hpp"
@@ -1216,7 +1216,7 @@ static void populate_partial_reduce_rt_args(
 
     auto get_fabric_connection = [&device, &fabric](bool is_connected, Direction dir) {
         return is_connected
-                   ? std::make_optional<ttnn::ccl::SenderWorkerAdapterSpec>(fabric.uniquely_connect_worker(device, dir))
+                   ? std::make_optional<tt::fabric::SenderWorkerAdapterSpec>(fabric.uniquely_connect_worker(device, dir))
                    : std::nullopt;
     };
 
@@ -1540,7 +1540,7 @@ static void create_end_of_line_worker_runtime_args(
 
     auto get_fabric_connection = [&device, &fabric](bool is_connected, Direction dir) {
         return is_connected
-                   ? std::make_optional<ttnn::ccl::SenderWorkerAdapterSpec>(fabric.uniquely_connect_worker(device, dir))
+                   ? std::make_optional<tt::fabric::SenderWorkerAdapterSpec>(fabric.uniquely_connect_worker(device, dir))
                    : std::nullopt;
     };
